Fix bugs on invalid inputs
