+++ data/test.txt
@@ -3,3 +3,11 @@ E & 0 &  1 & meeting & 21/10/2021 1200
 T & 0 &  1 & buy milk
 E & 0 &  1 & meeting & 21/10/2021 1200
 T & 0 &  1 & buy milk
+E & 0 &  1 & meeting & 21/10/2021 1200
+T & 0 &  1 & buy milk
+E & 0 &  1 & meeting & 21/10/2021 1200
+T & 0 &  1 & buy milk
+E & 0 &  1 & meeting & 21/10/2021 1200
+T & 0 &  1 & buy milk
+E & 0 &  1 & meeting & 21/10/2021 1200
+T & 0 &  1 & buy milk
